refactor: serviceImpl 수정
