+++ src/main/java/com/sparta/backend/utils/DummyRequestDto.java
@@ -1,0 +1,12 @@
+package com.sparta.backend.utils;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class DummyRequestDto {
+    private int users;
+    private int articleFolders;
+    private int articles;
+}
